% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowFunc.R
\name{bow}
\alias{bow}
\title{bow}
\usage{
bow(s, n)
}
\arguments{
\item{s}{is the string or fasta sequence, of which you want the Bow representation}

\item{n}{is the length of the kmers}
}
\value{
a dataframe, representing the Bag of words of your string
}
\description{
bow
}
