% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowFunc.R
\name{bow_show}
\alias{bow_show}
\title{bow show}
\usage{
bow_show(Bow)
}
\arguments{
\item{Bow}{the dataframe returned by the function bow}
}
\value{
a barplot with the representation of your k-mer with respect to the mean.
}
\description{
bow show
}
