% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowFunc.R
\name{get_kmer_all}
\alias{get_kmer_all}
\title{get_kmer_all}
\usage{
get_kmer_all(x, n)
}
\arguments{
\item{x}{string}

\item{n}{length of the kmer}
}
\value{
list of kmer with repetition of each kmer.
}
\description{
get_kmer_all
}
